48. Rotate Image
